[BUGFIX] Line endings in ICS must be Windows style
According to RFC2445 (iCalendar specification), line endings must be in
Windows style (\r\n). The TrimViewHelper is adapted to replace other
line endings to Windows style for ICS only.

Furthermore, line feeds in the DESCRIPTION (teaser field of news) must
be processed to have a space at the beginning of a new line to make the
feed validate.

Resolves: #61731
Releases: 3.1
Change-Id: I1a4cdee123a5d6074f6c43330f23c7bf18d5ab19
Reviewed-on: http://review.typo3.org/37673
Reviewed-by: Patrick Wiggelman <patrickwiggelman@roquin.nl>
Tested-by: Patrick Wiggelman <patrickwiggelman@roquin.nl>