[BUGFIX] Active constraint must consider one-day events without time
If you have a one-day event without a time, normally you just enter a
start date, but neither a start time nor an end date nor an end time.

That way the event won't be active the whole day because for
roq_newsevent the event starts and ends at 0:00:00 of the day.

To also consider such events, an additional constraint "current single
day events without time" should be added.

Resolves: #60291
Releases: 3.1
Change-Id: Ibb3686ac1bd1946fe2dee7b5844e83919ff49a9e
Reviewed-on: http://review.typo3.org/37674
Reviewed-by: Patrick Wiggelman <patrickwiggelman@roquin.nl>
Tested-by: Patrick Wiggelman <patrickwiggelman@roquin.nl>